requirements.txt
python-dotenv
langchain-openai
langchain-community
langchain-core
streamlit
requests
fastapi
pydantic
chromadb
